added ver1.2 copy from megu, archived older version
